+++ lib/features/profile/widgets/button_widget.dart
@@ -2,7 +2,7 @@
 Button widget is used by multiple pages for operations like submit, log in, log out etc
 */
 
-import "package:flutter/material.dart";
+import 'package:flutter/material.dart';
 
 class ButtonWidget extends StatelessWidget {
   final String text;
@@ -17,9 +17,9 @@ class ButtonWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) => ElevatedButton(
         style: ElevatedButton.styleFrom(
-          shape: StadiumBorder(),
+          shape: const StadiumBorder(),
           onPrimary: Colors.white,
-          padding: EdgeInsets.symmetric(horizontal: 32, vertical: 12),
+          padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 12),
         ),
         child: Text(text),
         onPressed: onClicked,
