+++ lib/features/profile/widgets/profile_widget.dart
@@ -19,16 +19,19 @@ class ProfileWidget extends StatelessWidget {
     final color = Theme.of(context).colorScheme.primary;
 
     return Center(
-        child: Stack(children: [
-      buildImage(),
+      child: Stack(
+        children: [
+          buildImage(),
 
-      //for editing emails and passwords
-      Positioned(
-        bottom: 0,
-        right: 4,
-        child: buildEditIcon(color),
+          //for editing emails and passwords
+          Positioned(
+            bottom: 0,
+            right: 4,
+            child: buildEditIcon(color),
+          )
+        ]
       )
-    ]));
+    );
   }
 
   Widget buildImage() {
@@ -49,30 +52,32 @@ class ProfileWidget extends StatelessWidget {
   }
 
   Widget buildEditIcon(Color color) => buildCircle(
+    color: Colors.white,
+    all: 3,
+    child: buildCircle(
+      color: color,
+      all: 8,
+      child: Icon(
+        // Change icons according to which mode edit or profile page
+        isEdit ? Icons.add_a_photo : Icons.edit,
         color: Colors.white,
-        all: 3,
-        child: buildCircle(
-          color: color,
-          all: 8,
-          child: Icon(
-            // Change icons according to which mode edit or profile page
-            isEdit ? Icons.add_a_photo : Icons.edit,
-            color: Colors.white,
-            size: 30,
-          ),
-          
-        ),
-      );
+        size: 30,
+      ),
+      
+    ),
+  );
 
   Widget buildCircle({
     required Widget child,
     required double all,
     required Color color,
-  }) =>
-      ClipOval(
-          child: Container(
+  }) {
+    return ClipOval(
+      child: Container(
         padding: EdgeInsets.all(all),
         color: color,
         child: child,
-      ));
+      ),
+    );
+  }
 }
